chore #19 : 라우트 세팅 및 몇몇 page 지정
