+++ package.json
@@ -19,6 +19,7 @@
     "prettier": "^3.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
