+++ src/page/Auth.tsx
@@ -1,0 +1,14 @@
+import { styled } from 'styled-components';
+
+export function Auth() {
+  return <Div>로그인/회원가입/프로필</Div>;
+}
+
+const Div = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`;
